+++ packages/spacecat-shared-ahrefs-client/CHANGELOG.md
@@ -1,3 +1,10 @@
+# [@adobe/spacecat-shared-ahrefs-client-v1.9.7](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-ahrefs-client-v1.9.6...@adobe/spacecat-shared-ahrefs-client-v1.9.7) (2025-09-09)
+
+
+### Bug Fixes
+
+* **deps:** update external major (major) ([#795](https://github.com/adobe/spacecat-shared/issues/795)) ([b020e88](https://github.com/adobe/spacecat-shared/commit/b020e884bfcad48667da87ad9caee7a3669e43d0))
+
 # [@adobe/spacecat-shared-ahrefs-client-v1.9.6](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-ahrefs-client-v1.9.5...@adobe/spacecat-shared-ahrefs-client-v1.9.6) (2025-09-06)